[#140] rename: 컬렉터 디렉터리 구조 변경
