+++ collector/src/main/java/com/wherecar/collector/application/GpsLogConverterService.java
@@ -1,0 +1,9 @@
+package com.wherecar.collector.application;
+
+import com.wherecar.collector.application.dto.GpsLogRequest;
+import com.wherecar.collector.application.dto.GpsLogResponse;
+
+public interface GpsLogConverterService {
+
+    GpsLogResponse receiveGpsLog(GpsLogRequest gpsLogRequest);
+}
