+++ collector/src/main/java/com/wherecar/collector/application/dto/GpsLogInfo.java
@@ -1,4 +1,4 @@
-package com.wherecar.collector.dto;
+package com.wherecar.collector.application.dto;
 
 import lombok.*;
