+++ collector/src/main/java/com/wherecar/collector/application/dto/GpsLogResponse.java
@@ -1,4 +1,4 @@
-package com.wherecar.collector.dto;
+package com.wherecar.collector.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
