+++ collector/src/main/java/com/wherecar/collector/infrastructure/GpsLogRepository.java
@@ -1,4 +1,4 @@
-package com.wherecar.collector.repository;
+package com.wherecar.collector.infrastructure;
 
 import com.wherecar.collector.domain.GpsLog;
 import org.springframework.data.jpa.repository.JpaRepository;
