--- collector/src/main/java/com/wherecar/collector/service/GpsLogConverterService.java
@@ -1,9 +1,0 @@
-package com.wherecar.collector.service;
-
-import com.wherecar.collector.dto.GpsLogRequest;
-import com.wherecar.collector.dto.GpsLogResponse;
-
-public interface GpsLogConverterService {
-
-    GpsLogResponse receiveGpsLog(GpsLogRequest gpsLogRequest);
-}
